+++ pytest_helm_charts/flux/helm_release.py
@@ -5,7 +5,7 @@ from typing import Protocol, Optional, List, Any, Dict
 from pykube import HTTPClient
 
 from pytest_helm_charts.k8s.fixtures import NamespaceFactoryFunc
-from pytest_helm_charts.flux.utils import NamespacedFluxCR, FLUX_CR_READY_TIMEOUT_SEC, _flux_cr_ready
+from pytest_helm_charts.flux.utils import NamespacedFluxCR, FLUX_CR_READY_TIMEOUT_SEC, flux_cr_ready
 from pytest_helm_charts.utils import wait_for_objects_condition, inject_extra
 
 
@@ -206,7 +206,7 @@ def wait_for_helm_releases_to_be_ready(
         HelmReleaseCR,
         helm_release_names,
         helm_release_namespace,
-        _flux_cr_ready,
+        flux_cr_ready,
         timeout_sec,
         missing_ok,
     )