+++ pytest_helm_charts/flux/kustomization.py
@@ -4,7 +4,7 @@ from typing import Protocol, Optional, Any, List, Dict
 from pykube import HTTPClient
 
 from pytest_helm_charts.k8s.fixtures import NamespaceFactoryFunc
-from pytest_helm_charts.flux.utils import NamespacedFluxCR, FLUX_CR_READY_TIMEOUT_SEC, _flux_cr_ready
+from pytest_helm_charts.flux.utils import NamespacedFluxCR, FLUX_CR_READY_TIMEOUT_SEC, flux_cr_ready
 from pytest_helm_charts.utils import wait_for_objects_condition, inject_extra
 
 
@@ -149,7 +149,7 @@ def wait_for_kustomizations_to_be_ready(
         KustomizationCR,
         kustomization_names,
         kustomization_namespace,
-        _flux_cr_ready,
+        flux_cr_ready,
         timeout_sec,
         missing_ok,
     )